+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 1.0.0
 
+- Support to spawn Dart entry points: script, file and Uri.
+  - Allow the use of another project/package dependencies while
+    running a spawned Dart entry point.
 - Initial version.